Some more work for the base accessibility implementation.
- size_t/ssize_t instead of int + show menu in Accessible.
- MFDrawBox now returns the proper accessibility values (title, identifier, label, value etc.) and can show a menu.
- Home screen sidebar, tabs, welcome + connection screens fully support accessibility now.
- Some cleanup + fixed warnings.

+++ frontend/common/snippet_list.cpp
@@ -580,12 +580,14 @@ base::Rect BaseSnippetList::snippet_bounds(Snippet* snippet) {
 }
 
 //------------------------------------------------------------------------------------------------
-int BaseSnippetList::getAccessibilityChildCount() {
+
+size_t BaseSnippetList::getAccessibilityChildCount() {
   return (int)_snippets.size();
 }
 
 //------------------------------------------------------------------------------------------------
-Accessible* BaseSnippetList::getAccessibilityChild(int index) {
+
+Accessible* BaseSnippetList::getAccessibilityChild(size_t index) {
   base::Accessible* accessible = NULL;
 
   if ((size_t)index < _snippets.size())
@@ -595,6 +597,7 @@ Accessible* BaseSnippetList::getAccessibilityChild(int index) {
 }
 
 //------------------------------------------------------------------------------------------------
-base::Accessible* BaseSnippetList::accessibilityHitTest(int x, int y) {
+
+base::Accessible* BaseSnippetList::accessibilityHitTest(ssize_t x, ssize_t y) {
   return snippet_from_point(x, y);
 }
