+++ library/forms/tabswitcher.cpp
@@ -634,14 +634,14 @@ bool TabSwitcher::mouse_leave() {
 
 //--------------------------------------------------------------------------------------------------
 
-int TabSwitcher::getAccessibilityChildCount() {
+size_t TabSwitcher::getAccessibilityChildCount() {
   return _pimpl->getItemCount();
 }
 
 //--------------------------------------------------------------------------------------------------
 
-Accessible *TabSwitcher::getAccessibilityChild(int index) {
-  return dynamic_cast<Accessible*>(_pimpl->getItem(index));
+Accessible *TabSwitcher::getAccessibilityChild(size_t index) {
+  return dynamic_cast<Accessible*>(_pimpl->getItem(static_cast<int>(index)));
 }
 
 //--------------------------------------------------------------------------------------------------
@@ -652,8 +652,8 @@ Accessible::Role TabSwitcher::getAccessibilityRole() {
 
 //--------------------------------------------------------------------------------------------------
 
-base::Accessible *TabSwitcher::accessibilityHitTest(int x, int y) {
-  int idx = _pimpl->index_from_point(x, y);
+base::Accessible *TabSwitcher::accessibilityHitTest(ssize_t x, ssize_t y) {
+  int idx = _pimpl->index_from_point(static_cast<int>(x), static_cast<int>(y));
   if (idx == -1)
     return nullptr;
   else
